Runlings on vampire playing cards "as if"
This is a recurring question and a dozen vampires are concerned.

For, now , only the latest ruling has been added on the question,
with wording coming from the rulebook
(from allies playing cards "as if" they were vampires), as it works the same.
The discussion is quite long around it
and it lists a bunch of other rulings that where not referenced.