Improve wording: remove fizzle and impulse
Neither of those terms are in the rulebook,
even if they are often used in rulings.

A fizzled action has no effect but  is considered successful,
which is not immediately intuitive.

Impulse is practical but a player won't find the definition in the rulebook
and may wonder what it means.

An alternative would be additional data structures and endpoints,
for example "glossary" or "definitions" provided along the rulings.

For now let's try to KISS,
even if this makes the rulings a bit more circumvoluted.